Split spead2::send::writer into its own files
