+++ include/spead2/send_streambuf.h
@@ -25,6 +25,7 @@
 #include <functional>
 #include <boost/asio.hpp>
 #include <spead2/send_stream.h>
+#include <spead2/send_writer.h>
 
 namespace spead2
 {
