+++ src/Makefile.am
@@ -87,4 +87,5 @@ libspead2_a_SOURCES = \
 	send_stream.cpp \
 	send_tcp.cpp \
 	send_udp.cpp \
-	send_udp_ibv.cpp
+	send_udp_ibv.cpp \
+	send_writer.cpp
